Bump to 0.0.3: remove postinstall script
Postinstall contained build script which requires tsc when installing from NPM.
It is actually not needed because ts files are compiled before publishing with prepublish script.
Postinstall was intended to be used when installing from git, but its inconveniences
outweigh the benefits. After all, installing by `npm run build` is not that difficult.